TimeTracking API: New Project Endpoints & Refactoring
General GET and CRUD endpoints removed.
New endpoints added for time breakdown and total duration per project.
Start/Stop now check task assignment.
Duration changed from double to int.
Additional console output added for error diagnosis.

+++ Models/TimeTracking.cs
@@ -29,7 +29,7 @@ namespace SmartTaskTracker.API.Models
         public DateTime? EndTime { get; set; }
 
         [Column("duration")]
-        public double? Duration { get; set; } 
+        public int? Duration { get; set; } 
         
     }
 }
